Refactor: Modularize Options class
Extract utility classes for better organization and maintainability.
